+++ bindings/nodejs/types/bridge/index.ts
@@ -22,8 +22,9 @@ import type {
     __ListPendingTransactionsMethod__,
     __ListTransactionsMethod__,
     __ListUnspentOutputsMethod__,
-    __MeltNativeTokenMethod__,
+    __DecreaseNativeTokenSupplyMethod__,
     __MinimumRequiredStorageDepositMethod__,
+    __IncreaseNativeTokenSupplyMethod__,
     __MintNativeTokenMethod__,
     __MintNftsMethod__,
     __PrepareOutputMethod__,
@@ -90,8 +91,9 @@ export type __AccountMethod__ =
     | __ListPendingTransactionsMethod__
     | __ListTransactionsMethod__
     | __ListUnspentOutputsMethod__
-    | __MeltNativeTokenMethod__
+    | __DecreaseNativeTokenSupplyMethod__
     | __MinimumRequiredStorageDepositMethod__
+    | __IncreaseNativeTokenSupplyMethod__
     | __MintNativeTokenMethod__
     | __MintNftsMethod__
     | __PrepareOutputMethod__