+++ bindings/python/native/iota_wallet/account.py
@@ -115,22 +115,24 @@ class Account:
         )
 
     def burn_native_token(self,
-                          native_token,
-                          options):
+                          token_id,
+                          burn_amount,
+                          options=None):
         """Burn native tokens. This doesn't require the foundry output which minted them, but will not increase
         the foundries `melted_tokens` field, which makes it impossible to destroy the foundry output. Therefore it's
         recommended to use melting, if the foundry output is available.
         """
         return self._call_account_method(
             'BurnNativeToken', {
-                'nativeToken': native_token,
+                'tokenId': token_id,
+                'burnAmount': burn_amount,
                 'options':  options
             }
         )
 
     def burn_nft(self,
                  nft_id,
-                 options):
+                 options=None):
         """Burn an nft output. Outputs controlled by it will be sweeped before if they don't have a storage
         deposit return, timelock or expiration unlock condition. This should be preferred over burning, because after
         burning, the foundry can never be destroyed anymore.
@@ -142,9 +144,9 @@ class Account:
             }
         )
 
-    def cnsolidate_outputs(self,
-                           force,
-                           output_consolidation_threshold):
+    def consolidate_outputs(self,
+                            force,
+                            output_consolidation_threshold):
         """Consolidate outputs.
         """
         return self._call_account_method(
@@ -156,7 +158,7 @@ class Account:
 
     def destroy_alias(self,
                       alias_id,
-                      options):
+                      options=None):
         """Destroy an alias output. Outputs controlled by it will be sweeped before if they don't have a
         storage deposit return, timelock or expiration unlock condition. The amount and possible native tokens will be
         sent to the governor address.
@@ -170,7 +172,7 @@ class Account:
 
     def destroy_foundry(self,
                         foundry_id,
-                        options):
+                        options=None):
         """Destroy a foundry output with a circulating supply of 0.
         Native tokens in the foundry (minted by other foundries) will be transactioned to the controlling alias
         """
@@ -232,18 +234,22 @@ class Account:
             'ListAddressesWithUnspentOutputs'
         )
 
-    def list_outputs(self):
+    def list_outputs(self, filter_options=None):
         """Returns all outputs of the account.
         """
         return self._call_account_method(
-            'ListOutputs'
+            'ListOutputs', {
+                'filterOptions': filter_options
+            }
         )
 
-    def list_unspent_outputs(self):
+    def list_unspent_outputs(self, filter_options=None):
         """Returns all unspent outputs of the account.
         """
         return self._call_account_method(
-            'ListUnspentOutputs'
+            'ListUnspentOutputs', {
+                'filterOptions': filter_options
+            }
         )
 
     def list_transactions(self):
@@ -260,20 +266,34 @@ class Account:
             'ListPendingTransactions'
         )
 
-    def melt_native_token(self,
-                          native_token,
-                          options):
+    def decrease_native_token_supply(self,
+                                     token_id,
+                                     melt_amount,
+                                     options=None):
         """Melt native tokens. This happens with the foundry output which minted them, by increasing it's
         `melted_tokens` field.
         """
         return self._call_account_method(
-            'MeltNativeToken', {
-                'nativeToken': native_token,
+            'DecreaseNativeTokenSupply', {
+                'tokenId': token_id,
+                'meltAmount': melt_amount,
                 'options':  options
             }
         )
 
-    def mint_native_token(self, native_token_options, options):
+    def increase_native_token_supply(self, token_id, mint_amount, increase_native_token_supply_options=None, options=None):
+        """Mint more native token.
+        """
+        return self._call_account_method(
+            'IncreaseNativeTokenSupply', {
+                'tokenId': token_id,
+                'mintAmount': mint_amount,
+                'increaseNativeTokenSupplyOptions': increase_native_token_supply_options,
+                'options': options
+            }
+        )
+
+    def mint_native_token(self, native_token_options, options=None):
         """Mint native token.
         """
         return self._call_account_method(
@@ -292,7 +312,7 @@ class Account:
             }
         )
 
-    def mint_nfts(self, nfts_options, options):
+    def mint_nfts(self, nfts_options, options=None):
         """Mint nfts.
         """
         return self._call_account_method(
@@ -309,7 +329,7 @@ class Account:
             'GetBalance'
         )
 
-    def prepare_send_amount(self, addresses_with_amount, options):
+    def prepare_send_amount(self, addresses_with_amount, options=None):
         """Prepare send amount.
         """
         return self._call_account_method(
@@ -319,7 +339,7 @@ class Account:
             }
         )
 
-    def prepare_transaction(self, outputs, options):
+    def prepare_transaction(self, outputs, options=None):
         """Prepare transaction.
         """
         return self._call_account_method(
@@ -349,7 +369,7 @@ class Account:
             }
         )
 
-    def send_micro_transaction(self, addresses_with_micro_amount, options):
+    def send_micro_transaction(self, addresses_with_micro_amount, options=None):
         """Send micro transaction.
         """
         return self._call_account_method(
@@ -359,7 +379,7 @@ class Account:
             }
         )
 
-    def send_native_tokens(self, addresses_native_tokens, options):
+    def send_native_tokens(self, addresses_native_tokens, options=None):
         """Send native tokens.
         """
         return self._call_account_method(
@@ -369,7 +389,7 @@ class Account:
             }
         )
 
-    def send_nft(self, addresses_nft_ids, options):
+    def send_nft(self, addresses_nft_ids, options=None):
         """Send nft.
         """
         return self._call_account_method(
@@ -389,7 +409,7 @@ class Account:
             }
         )
 
-    def send_transaction(self, outputs, options):
+    def send_transaction(self, outputs, options=None):
         """Send a transaction.
         """
         return self._call_account_method(