+++ tests/burn_outputs.rs
@@ -39,7 +39,7 @@ async fn mint_and_burn_nft() -> Result<()> {
 
     let account = match manager.get_account("Alice".to_string()).await {
         Ok(account) => account,
-        Err(Error::AccountNotFound) => manager
+        Err(Error::AccountNotFound(_)) => manager
             .create_account()
             .with_alias("Alice".to_string())
             .finish()
@@ -89,7 +89,7 @@ async fn mint_and_burn_nft() -> Result<()> {
 
 #[ignore]
 #[tokio::test]
-async fn mint_and_melt_native_token() -> Result<()> {
+async fn mint_and_decrease_native_token_supply() -> Result<()> {
     let storage_path = "test-storage/mint_and_burn_outputs";
     std::fs::remove_dir_all(storage_path).unwrap_or(());
     let client_options = ClientOptions::new()
@@ -112,7 +112,7 @@ async fn mint_and_melt_native_token() -> Result<()> {
 
     let account = match manager.get_account("Alice".to_string()).await {
         Ok(account) => account,
-        Err(Error::AccountNotFound) => manager
+        Err(Error::AccountNotFound(_)) => manager
             .create_account()
             .with_alias("Alice".to_string())
             .finish()
@@ -153,7 +153,7 @@ async fn mint_and_melt_native_token() -> Result<()> {
     // Burn some of the circulating supply
     let burn_amount = U256::from(40i32);
     let _ = account
-        .melt_native_token((transaction.token_id, burn_amount), None)
+        .decrease_native_token_supply(transaction.token_id, burn_amount, None)
         .await
         .unwrap();
     tokio::time::sleep(Duration::new(15, 0)).await;
@@ -165,9 +165,9 @@ async fn mint_and_melt_native_token() -> Result<()> {
     assert!(search.is_some());
 
     // The burn the rest of the supply
-    let burn_amount = circulating_supply - burn_amount;
+    let melt_amount = circulating_supply - burn_amount;
     let _ = account
-        .melt_native_token((transaction.token_id, burn_amount), None)
+        .decrease_native_token_supply(transaction.token_id, melt_amount, None)
         .await
         .unwrap();
     tokio::time::sleep(Duration::new(15, 0)).await;
@@ -209,7 +209,7 @@ async fn destroy_foundry() -> Result<()> {
 
     let account = match manager.get_account("Alice".to_string()).await {
         Ok(account) => account,
-        Err(Error::AccountNotFound) => manager
+        Err(Error::AccountNotFound(_)) => manager
             .create_account()
             .with_alias("Alice".to_string())
             .finish()